Started Komodo Project to Build a State Machine XML Parser for the Game
I started a Komodo Project to build a State Machine
parser for the Game Engine.  It will run off of two
controllers, game_controller and level_controller.
Both of these controllers should switch states based
on what the current tag is.

Headers:
-TagParse.h - Interface
-LevelParse.h
-NameParse.h
Source Files
-LevelParse.cpp
-NameParse.cpp

Also created a Trimmer.h and Trimmer.cpp to define
a function to trim whitespace off of strings passed
into it.  The function is defined in Trimmer.h and
implemented in Trimmer.cpp.  The function is known
as trim_ws.  It takes a G++ Standard string object as
the parameter and returns a G++ Standard string
object.