Implemented helper interfaces.
Comparable is an interface which forces its derived
classes to implement a compare function for its
templated object.

Pair.h is an interface which stores 2 objects.  It forces
its derived classes to implement a compare function
for both types of objects.  Written because std::pair
does not play nice with vectors.

PairException.h/.cpp is the exception handler for
Pair.h.