Defined and implemented core class in Core.h/.cpp
Defined core class in Core.h and implemented it in Core.cpp.

This class will be replacing the manager class defined earlier in the project.
It is built with the Mediator Design Pattern in mind.  The implementation of
manager contained dependency issues with render class.  The dependencies
were the scene class and level class.  I designed the class core with the intention
to eliminate these dependencies as well as to enforce myself to document the
methods which are handling the message passing and minimize the number
of methods needing to be defined, since, script manager has now been created
and many of the predefined methods can now be moved into script calls.